LPS-55917 Move
